Assignment_4/requirements.txt
absl-py==0.7.0
appnope==0.1.0
astor==0.7.1
autopep8==1.4.3
backcall==0.1.0
bleach==2.1.3
cycler==0.10.0
decorator==4.3.0
entrypoints==0.2.3
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
html5lib==1.0.1
ipdb==0.11
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.2
mistune==0.8.3
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
numpy==1.14.3
pandas==0.23.4
pandocfilters==1.4.2
parso==0.2.0
pbr==5.1.3
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.4.1
prompt-toolkit==1.0.15
protobuf==3.7.0
ptyprocess==0.5.2
pycodestyle==2.5.0
Pygments==2.2.0
pyparsing==2.3.1
python-dateutil==2.7.3
pytz==2018.5
pyzmq==17.0.0
qtconsole==4.3.1
scikit-learn==0.20.2
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
